Using the Sked icon for the main window and password dialog
This is not the best way to do it. There are too many data files
on the libsked directory which should be moved to /usr/share/sked/,
but it avoids hardcoded paths, works when the program is ran from
any directory or without installation, enables the Debian packaging
with python-support, and allows 'setup.py' to put the files in the
right place on Windows too. On Unix builds, it also has the drawback
of duplicating 'sked.png'.