Make HistoryManager independent from SkedApp and 'max_items' optional
The class HistoryManager does not depends on SkedApp anymore, so now
it may be used in non-interactive processes. The constructor now
receives a database object. The maximum length of the history is now
given in a optional property, instead of being loaded from an
OptionManager instance.

This commit also adds the methods 'set_items' and 'load' to allow
resetting the list of items and (re)loading its contents from the
database.