Add selective import/export of XML data
This commit adds options to select what kind of information will be
exported to or imported from the XML data files. For example, it allows
exporting only the pages of a database (option "Sked XML, pages only"),
so the resulting file can be imported in older versions of Sked.

It also moves the relevant code to its own class, easing the future
implementation of more features.