Make the handling of formatting tags more clever
The commands to add formatting codes now toggles the tags in the
borders of the selection. The commands to add section headers also
are a bit more clever now, removing other header tags if necessary.

Also, when format commands are issued with some text selected, the
cursor is moved to the end of the selection, and the selection is
removed. If no text is selected, the cursor is placed between the
format tags. This fixes a very old bug.