Add a "clipboard contents" ("\c") token to the macro interpreter
The clipboard may have some large chunks of data selected, so it makes
a lot of sense only evaluate it when needed. The token_dict was changed
to accommodate this new requirement and it now accepts callable objects
as values too. These objects are called when needed and its return
values are used if applicable. Also, as the backslash notation parser
is getting more and more complex, the strftime tokens are now evaluated
first.