Allow customization of automatic pages with templates
Add a new feature to allow the customization of the default text added
when the user follows a link to a non-existing page and the optional
redirection page created by the command 'Rename page'.

These templates are stored in the options and evaluated through the
macro system when needed, so, all macro tokens are available to the
templates too.